Changed some data types

+++ AnimeBytes-Helper/functions.h
@@ -1,11 +1,11 @@
 #pragma once
 
 void end();
-float getValueFromUser();
-int calcSeedTime(float gigabytes);
-std::string days(int d, short h);
+double getValueFromUser();
+int calcSeedTime(double gigabytes);
+std::string days(int d, int h);
 void formatAndPrintSeedTime(int seedTime);
 void formatAndPrintGracePeriod(int gracePeriod);
 
-// --- AnimeBytes Helper Functions ---
+/* --- AnimeBytes Helper Functions --- */
 // void seedingTimeCalculator();
